pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.3.0",
    "plotly>=6.1.2",
    "streamlit-drawable-canvas>=0.9.3",
    "streamlit>=1.46.0",
    "trafilatura>=2.0.0",
    "yfinance>=0.2.63",
    "pillow>=11.2.1",
    "kaleido>=1.0.0",
    "numpy>=2.3.1",
    "matplotlib>=3.10.3",
]
